LibGUI: Store NumericInput value as an i64
Moar range = moar better. I'm hoping to use this for a size_t value
later, so this almost covers that.

+++ Userland/Libraries/LibGUI/NumericInput.h
@@ -16,20 +16,20 @@ class NumericInput final : public GUI::TextBox {
 public:
     virtual ~NumericInput() override = default;
 
-    Function<void(i32)> on_number_changed;
+    Function<void(i64)> on_number_changed;
 
-    void set_min_number(i32 number);
-    void set_max_number(i32 number);
-    void set_current_number(i32 number, GUI::AllowCallback allow_callback = GUI::AllowCallback::Yes);
+    void set_min_number(i64 number);
+    void set_max_number(i64 number);
+    void set_current_number(i64 number, GUI::AllowCallback allow_callback = GUI::AllowCallback::Yes);
 
 private:
     NumericInput();
     void on_focus_lost();
 
     bool m_needs_text_reset { false };
-    i32 m_current_number { 0 };
-    i32 m_min_number { NumericLimits<i32>::min() };
-    i32 m_max_number { NumericLimits<i32>::max() };
+    i64 m_current_number { 0 };
+    i64 m_min_number { NumericLimits<i64>::min() };
+    i64 m_max_number { NumericLimits<i64>::max() };
 };
 
 }
